Removed (broken) email invitation functionality
The email invitation functionality is broken.  After looking at this
code, which was written by contractors, I
have come to the conclusion that it is extremely flaky, unmaintainable,
and almost unreadable.  There are no tests, things return null pointers
all over the place, exceptions are caught and squashed etc. etc.

I think it is therefore best to kill this off and reimplement it from
scratch.  Removing this allows loads of other dodgy code
(AccessPolicyService
in particular) to be removed.  It also helps us on our way to
eradicating gwtext.

There is a workaround for not having email invitation documented by
Tania.  See issue #232.  I think this is perfectly satisfactory for the
moment.